+++ src/FocusApp/App.cs
@@ -1,9 +1,4 @@
 ﻿using FocusApp.Resources;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FocusApp
 {
@@ -13,7 +8,7 @@ namespace FocusApp
         {
             Resources = new AppStyles();
 
-            MainPage = new AppShell();
+            MainPage = new Shell() { CurrentItem = new MainPage() };
         }
     }
 }
