--- src/FocusApp/AppShell.cs
@@ -1,12 +1,0 @@
-﻿using FocusApp.Pages;
-
-namespace FocusApp;
-
-class AppShell : Shell
-{
-
-	public AppShell()
-	{
-		Items.Add(new MainPage());	
-	}
-}
