+++ src/FocusApp/FocusApp.csproj
@@ -71,6 +71,7 @@
 
 	<ItemGroup>
 	  <Folder Include="ViewModels\" />
+	  <Folder Include="Views\" />
 	</ItemGroup>
 
 </Project>
